Move files to resourceTableComponent folder
