+++ src/constants.js
@@ -34,11 +34,6 @@ export const DEPLOYMENT_STATUS = {
   ERROR: "ERROR",
 };
 
-export const RESOURCE_TABLE_TYPE = {
-  LAMBDA: "LAMBDA",
-  ALL_RESOURCES: "ALL_RESOURCES",
-};
-
 export const DASHBOARD_FOCUS_INDEX = {
   RESOURCE_TABLE: 0,
   EVENT_BRIDGE_TREE: 1,
